% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOption.R
\name{abbreviatedOptions}
\alias{abbreviatedOptions}
\title{abbreviatedOptions}
\usage{
abbreviatedOptions()
}
\value{
a character vector with full and abbreviated option names
}
\description{
Returns the short forms for all global options.
}
\examples{
abbreviatedOptions()
}
