% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ID.R
\name{add_ID}
\alias{add_ID}
\title{add_ID}
\usage{
add_ID(id, txt = "")
}
\arguments{
\item{id}{character: name of ID}

\item{txt}{character: additional text informations to the ID (default: '')}
}
\description{
Adds the ID (name of an input or output element of the app) to the set of IDs. If the ID already exists then an error will be thrown.
}
\examples{
add_ID('test')
}
