% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectIn.R
\name{as.choices}
\alias{as.choices}
\title{sliderIn}
\usage{
as.choices(txt, inc = NULL)
}
\arguments{
\item{dragRange}{}
}
\description{
Constructs a slider widget to select a numeric value from a range. For the not documented parameters see \code{\link[shiny]{sliderInput}}
}
\examples{
selectIn('mean', 'Mean', -3, 3, 0, 0.1)
}
