% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxIn.R
\name{checkboxIn}
\alias{checkboxIn}
\title{checkboxIn}
\usage{
checkboxIn(inputId, label, value = FALSE, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}
}
\value{
an environment
}
\description{
Creates a checkbox element.
}
\examples{
checkboxIn('rug', 'Show observations')
}
