% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceLevelIn.R
\name{confidenceLevelIn}
\alias{confidenceLevelIn}
\title{confidenceLevelIn}
\usage{
confidenceLevelIn(inputId, label, min, max, value, step, round = FALSE,
  format = NULL, locale = NULL, ticks = TRUE, animate = FALSE,
  width = NULL, sep = ",", pre = NULL, post = NULL,
  timeFormat = NULL, timezone = NULL, dragRange = TRUE,
  percent = FALSE)
}
\arguments{
\item{inputId}{}

\item{min}{numeric: if \code{percent==TRUE}  \emph{80} else \emph{0.8}}

\item{max}{numeric:  if \code{percent==TRUE}  \emph{99.99} else \emph{0.999}}

\item{value}{numeric: if \code{percent==TRUE}  \emph{95} else \emph{0.95}}

\item{step}{numeric: if \code{percent==TRUE}  \emph{0.1} else \emph{0.001}}

\item{round}{}

\item{format}{}

\item{locale}{}

\item{ticks}{}

\item{animate}{}

\item{width}{}

\item{sep}{}

\item{pre}{}

\item{post}{}

\item{timeFormat}{}

\item{timezone}{}

\item{dragRange}{}

\item{percent}{logical: use 95 or 0.95 (default: FALSE)}
}
\description{
Constructs a slider to select a confidence level. For the undocumented parameters see \code{\link[shiny]{sliderInput}}
}
\examples{
confidenceLevelInput('cl')
}
