% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationIn.R
\name{correlationIn}
\alias{correlationIn}
\title{correlationIn}
\usage{
correlationIn(inputId, label = "Correlation (r)", min = -1, max = +1,
  value = 0, step = 0.01, round = FALSE, format = NULL,
  locale = NULL, ticks = TRUE, animate = FALSE, width = NULL,
  sep = ",", pre = NULL, post = NULL, timeFormat = NULL,
  timezone = NULL, dragRange = TRUE)
}
\arguments{
\item{inputId}{character: The input slot that will be used to access the value.}

\item{label}{character: Display label for the control, or \code{NULL} for no label.}

\item{min}{numeric: The minimum value (inclusive) that can be selected (default: 30)}

\item{max}{numeric: The maximum value (inclusive) that can be selected (default: 1000)}

\item{value}{numeric: The initial value of the slider}

\item{step}{numeric: Specifies the interval between each selectable value on the slider (if NULL, a heuristic is used to determine the step size, default: 10)}

\item{round}{}

\item{format}{}

\item{locale}{}

\item{ticks}{}

\item{animate}{}

\item{width}{}

\item{sep}{}

\item{pre}{}

\item{post}{}

\item{timeFormat}{}

\item{timezone}{}

\item{dragRange}{}
}
\description{
Constructs a slider widget to select a correlation value. For the not documented parameters see \code{\link[shiny]{sliderInput}}
}
\examples{
correlationIn('r', value=0)
}
