% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIn.R
\name{dataIn}
\alias{dataIn}
\title{dataIn}
\usage{
dataIn(inputId, label, ..., selected = NULL, selectize = TRUE,
  width = NULL, size = NULL, file = NULL)
}
\arguments{
\item{inputId}{character: name of data element}

\item{...}{further parameters to select one or more variables, for details see \code{\link{varIn}}}

\item{selected}{}

\item{selectize}{}

\item{width}{}

\item{size}{}

\item{file}{character: name of the RDS file with a list of data sets}
}
\description{
Constructs a selection to select a data sets. For the undocumented parameters see \code{\link[shiny]{selectInput}}
}
