% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{flatten}
\alias{flatten}
\title{flatten}
\usage{
flatten(lst, prefix = "", sep = "$")
}
\arguments{
\item{lst}{list: nested list}

\item{prefix}{character: prefix  (default: '')}

\item{sep}{character: separator for list elements (default: '$')}
}
\value{
simplified list
}
\description{
Flattens recursively a nested list to a simple list.
}
\examples{
flatten(list(a=1, b=list(c=2, d=3)))
}
