% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSelection.R
\name{getDataSelection}
\alias{getDataSelection}
\title{getDataSelection}
\usage{
getDataSelection(dindex, ..., simplify = TRUE)
}
\arguments{
\item{dindex}{numeric: index for data set}

\item{...}{numeric: index of variables in the data set, if not given the whole data frame is returned}

\item{simplify}{logical: simplify the set of selected variables (default: \code{TRUE})}
}
\value{
a vector, a data frame or a list of data frames depending on \code{simplify}
}
\description{
Returns a vector or data frame from the internally stored datasets in \code{mmstat}.
}
\examples{
\dontrun{getDataSelection(1)}
}
