% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getList.R
\name{getList}
\alias{getList}
\title{getList}
\usage{
getList(...)
}
\arguments{
\item{...}{named elements of a list}
}
\value{
renamed list
}
\description{
Creates a list with element names replaced by \code{link{getText}}.
}
\examples{
getList(BOSTON=1, MTCARS=2)
}
