% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMMstat.R
\name{getMMstat}
\alias{getMMstat}
\title{getMMstat}
\usage{
getMMstat(...)
}
\arguments{
\item{...}{elements}
}
\value{
the choosen element
}
\description{
Allows to access the package internal \code{mmstat} environment.
}
\examples{
getMMstat('version')
}
