% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOption.R
\name{getOption}
\alias{getOption}
\title{getOption}
\usage{
getOption(x, default = NULL)
}
\arguments{
\item{x}{a character string holding an option name.}

\item{default}{if the specified option is not set in the options list,
    this value is returned.  This facilitates retrieving an option and
    checking whether it is set and setting it separately if not.}
}
\value{
The current value set for option x, or NULL if the option is unset.
}
\description{
Allow the user to examine a variety of global options which affect the way in which R computes and displays its results.
An option of the form '\strong{m}mstat.\strong{c}ol.\strong{p}opulation' can be examined by the short form 'mcp' if unique.
}
\examples{
getOption('mmstat.col.population')
getOption('mcp')
}
