% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSelection.R
\name{getSelection}
\alias{getSelection}
\title{getSelection}
\usage{
getSelection(selectable, input, duplicates.ok = FALSE)
}
\arguments{
\item{selectable}{logical matrix: \code{TRUE} in row i and col j means that variable i in input j can be selected}

\item{input}{list of numeric vectors: list of wished input values}

\item{duplicates.ok}{logical: can variables be selected twice (default: \code{FALSE})}
}
\value{
a logical matrix with the entries 0 (var can not be selected), 1 (var can be selected) and 2 (var is selected)
}
\description{
Based on \code{selecttable} and a set of inputs
}
\examples{
sel <- cbind(c(FALSE,FALSE,TRUE,TRUE,FALSE), 
             c(TRUE,FALSE,TRUE,FALSE,TRUE), 
             c(TRUE,TRUE,FALSE,FALSE,TRUE))
getSelection(sel, list(1,1,1))
getSelection(sel, list(1,1,1), duplicates.ok=TRUE)
getSelection(sel, list(1,1,c(1,2,5)))
}
