% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSelection.R
\name{getSelectionAttr}
\alias{getSelectionAttr}
\title{getSelectionAttr}
\usage{
getSelectionAttr(dindex)
}
\arguments{
\item{dindex}{numeric: index of data set}
}
\value{
natrix
}
\description{
Returns the matrix attribute \code{selectable} which variables are selectable from data matrix in \code{mmstat}. If the attribute does not exists then 
a matrix is returned such that all variables are selectable
}
