% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getText.R
\name{getText}
\alias{getText}
\title{getText}
\usage{
getText(msg)
}
\arguments{
\item{msg}{character vector}
}
\value{
vector of translated messages
}
\description{
Translates a given message into another language.
}
\examples{
getText('Test')
}
