% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dev.R
\docType{data}
\name{golemLaunch}
\alias{golemLaunch}
\title{golemLaunch}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
golemLaunch
}
\value{
methods to manipulate your run_dev
}
\description{
A tools to use run_dev scripts
}
\section{Methods}{

\describe{
  \item{\code{show_terminal}}{Display the terminal where your application is launched}
  \item{\code{stop}}{Stop your app}
  \item{\code{get_buffer}}{Get what you have in your terminal}
  \item{\code{get_url}}{Get the url of your application}
  \item{\code{open_app}}{Open the url of your application}
  \item{\code{restart}}{Restart your application}
  \item{\code{is_running}}{Test your terminal is still running}
 }
}

\keyword{datasets}
