% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.true.R
\name{is.true}
\alias{is.true}
\alias{is.binary}
\title{Variable selection functions}
\usage{
is.true(x)

is.binary(x)
}
\arguments{
\item{x}{vector: variable values}
}
\value{
logical: if \code{x} fulfills the requirement
}
\description{
\itemize{
  \item \code{is.true} for any variable it returns \code{TRUE}, default at \code{\link{varIn}}
  \item \code{is.numeric} see \code{\link[base]{is.numeric}}
  \item \code{is.integer} see \code{\link[base]{is.integer}}
  \item \code{is.factor} see \code{\link[base]{is.factor}}
  \item \code{is.ordered} see \code{\link[base]{is.ordered}}
  \item \code{is.binary} returns \code{TRUE} if \code{x} takes only two values
}
}
\examples{
sapply(mtcars, is.true)
sapply(mtcars, is.binary)
}
