% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.libs.R
\name{list.libs}
\alias{list.libs}
\title{list.libs}
\usage{
list.libs(path = ".", pattern = "*.R$", all.files = FALSE,
  full.names = TRUE, recursive = TRUE, ignore.case = FALSE,
  include.dirs = FALSE, no.. = FALSE, quiet = FALSE)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link{getwd}()}.  Tilde
    expansion (see \code{\link{path.expand}}) is performed.  Missing
    values will be ignored.}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{all.files}{a logical value.  If \code{FALSE}, only the
    names of visible files are returned.  If \code{TRUE},
    all file names will be returned.}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{ignore.case}{logical.  Should pattern-matching be case-insensitive?}

\item{include.dirs}{logical.  Should subdirectory names be included in
    recursive listings?  (They always are in non-recursive ones).}

\item{no..}{logical.  Should both \code{"."} and \code{".."} be excluded
    also from non-recursive listings?}

\item{quiet}{logical: display the files analysed}
}
\value{
a table with \code{\link[base]{library}} and \code{\link[base]{require}} calls including frequencies
}
\description{
Collects all parameters \code{\link[base]{library}} and \code{\link[base]{require}} commands in a 
set of (R) files. In case that the parsing of a file fails a warning with the error message(s) will be 
displayed.
}
\examples{
list.libs(path=find.package("shinyExample"))
\dontrun{
  # install missing libraries
  eval(parse(text=attr(ll, "install"))) 
  # load all libraries
  eval(parse(text=attr(ll, "library")))
}
}
