% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShinyApp.R
\name{makeShinyApp}
\alias{makeShinyApp}
\title{makeShinyApp}
\usage{
makeShinyApp(output, input = NULL, dashboardHeader = list(title =
  "MM*Stat", titleWidth = NULL, disable = FALSE),
  dashboardSidebar = list(disable = FALSE, width = NULL, collapsed =
  FALSE), app = "app.tmpl")
}
\arguments{
\item{output}{list:}

\item{input}{list:}

\item{dashboardHeader}{list: parameters for \code{\link[shinydashboard]{dashboardHeader}} (default: \code{list(title=NULL, titleWidth=NULL, disable=FALSE)})}

\item{dashboardSidebar}{list: parameters for \code{\link[shinydashboard]{Sidebar}} (default: \code{list(disable=FALSE, width=NULL, collapsed=FALSE)})}

\item{app}{character: name of app template file (default: \code{app.tmpl})}
}
\description{
Creates a shiny app from one (or more) R example files using shinydashboard.
}
