% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutput.R
\name{plotOutput}
\alias{plotOutput}
\title{plotOutput}
\usage{
plotOutput(outputId, width = "100\%", height = "400px", click = NULL,
  dblclick = NULL, hover = NULL, hoverDelay = NULL,
  hoverDelayType = NULL, brush = NULL, clickId = NULL, hoverId = NULL,
  inline = FALSE, file = NULL, text = NULL)
}
\arguments{
\item{outputId}{}

\item{width}{}

\item{height}{}

\item{click}{}

\item{dblclick}{}

\item{hover}{}

\item{hoverDelay}{}

\item{hoverDelayType}{}

\item{brush}{}

\item{clickId}{}

\item{hoverId}{}

\item{inline}{}

\item{file}{character: name of file with R code to execute}

\item{text}{character: text with R code to execute}
}
\description{
Render a plot within an application page. For the not documented parameters see \code{\link[shiny]{plotOutput}}
}
\examples{
plotOutput('myplot', text='plot(1:10)')
\dontrun{
# create a static app
makeShinyApp(output=plotOutput('myplot', text='plot(1:10)'))
}
}
