% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyExample.R
\name{runShinyExample}
\alias{runShinyExample}
\title{runShinyExample}
\usage{
runShinyExample(launch.browser = getOption("shiny.launch.browser",
  interactive()))
}
\arguments{
\item{launch.browser}{logical (FALSE): If \code{TRUE} the app is launched in the system's default web browser.}
}
\value{
runs the shiny example app
}
\description{
starts a Shiny app that allows to run an R-, Octave- or Stata example programs
}
\examples{
\dontrun{
runShinyExample()
}
}
