% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSizeIn.R
\name{sampleSizeIn}
\alias{sampleSizeIn}
\title{sampleSizeIn}
\usage{
sampleSizeIn(inputId, label, min, max, value, step, round = TRUE,
  format = NULL, locale = NULL, ticks = TRUE, animate = FALSE,
  width = NULL, sep = ",", pre = NULL, post = NULL,
  timeFormat = NULL, timezone = NULL, dragRange = TRUE,
  log = FALSE)
}
\arguments{
\item{inputId}{character: The input slot that will be used to access the value.}

\item{label}{character: Display label for the control, or \code{NULL} for no label.}

\item{min}{numeric: The minimum value (inclusive) that can be selected (default: 30)}

\item{max}{numeric: The maximum value (inclusive) that can be selected (default: 1000)}

\item{value}{numeric: The initial value of the slider}

\item{step}{numeric: Specifies the interval between each selectable value on the slider (if NULL, a heuristic is used to determine the step size, default: 10)}

\item{round}{}

\item{format}{}

\item{locale}{}

\item{ticks}{}

\item{animate}{}

\item{width}{}

\item{sep}{}

\item{pre}{}

\item{post}{}

\item{timeFormat}{}

\item{timezone}{}

\item{dragRange}{}

\item{percent}{logical: use 95 or 0.95 (default: FALSE)}
}
\description{
Constructs a slider widget to select a sample size. For the not documented parameters see \code{\link[shiny]{sliderInput}}
}
\examples{
sampleSizeIn('n', value=50)
}
