% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectLanguage.R
\name{selectLanguage}
\alias{selectLanguage}
\title{selectLanguage}
\usage{
selectLanguage(index)
}
\arguments{
\item{index}{integer: number of language to use}
}
\value{
invisibly the selected language
}
\description{
Choose a language.
}
\examples{
selectLanguage(1)
}
