% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setData.R
\name{setData}
\alias{setData}
\title{setData}
\usage{
setData(...)
}
\arguments{
\item{...}{named list of data sets}
}
\value{
invisibly the names of all stored data sets
}
\description{
To store data sets in the library internal object \code{mmstat}.
}
