% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSelection.R
\name{setSelection}
\alias{setSelection}
\title{setSelection}
\usage{
setSelection(name, is)
}
\arguments{
\item{name}{character: name of variable}

\item{is}{function: function which returns for a vector \code{TRUE} or \code{FALSE}}
}
\value{
nothing
}
\description{
Set the \code{selectable} attribute of the internally stored data sets.
}
