% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyhist.R
\name{shinyhist}
\alias{shinyhist}
\title{This is an example of a shiny package}
\usage{
shinyhist()
}
\value{
A hstogram
}
\description{
This shows you where to place stuff
}
\details{
Get into reactive programming with shiny
}
\examples{
shinyhist()
}
