% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderIn.R
\name{sliderIn}
\alias{sliderIn}
\title{sliderIn}
\usage{
sliderIn(inputId, label, min, max, value, step = NULL, round = FALSE,
  format = NULL, locale = NULL, ticks = TRUE, animate = FALSE,
  width = NULL, sep = ",", pre = NULL, post = NULL,
  timeFormat = NULL, timezone = NULL, dragRange = TRUE)
}
\arguments{
\item{dragRange}{}
}
\description{
Constructs a slider widget to select a numeric value from a range. For the not documented parameters see \code{\link[shiny]{sliderInput}}
}
\examples{
sliderIn('mean', 'Mean', -3, 3, 0, 0.1)
}
