% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliderInput.R
\name{sliderInput}
\alias{sliderInput}
\title{sliderInput}
\usage{
sliderInput(inputId, label, min, max, value, step = NULL, round = FALSE,
  format = NULL, locale = NULL, ticks = TRUE, animate = FALSE,
  width = NULL, sep = ",", pre = NULL, post = NULL, timeFormat = NULL,
  timezone = NULL, dragRange = TRUE)
}
\arguments{
\item{inputId}{}

\item{label}{}

\item{min}{}

\item{max}{}

\item{value}{}

\item{step}{}

\item{round}{}

\item{format}{}

\item{locale}{}

\item{ticks}{}

\item{animate}{}

\item{width}{}

\item{sep}{}

\item{pre}{}

\item{post}{}

\item{timeFormat}{}

\item{timezone}{}

\item{dragRange}{}
}
\description{
Constructs a slider widget to select a numeric value from a range. For the not documented parameters see \code{\link[shiny]{sliderInput}}
}
\examples{
sliderInput('mean', 'Mean', -3, 3, 0, 0.1)
}
