% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_call.R
\name{str_call}
\alias{str_call}
\title{str_call}
\usage{
str_call(func, args, lib = NULL)
}
\arguments{
\item{func}{character: a non-empty character string naming the function to be called.}

\item{args}{list: a list of arguments to the function call. The names attribute of args gives the argument names.}

\item{lib}{character: name of library as prefix ot the function to be called (default: \code{NULL})}
}
\value{
a text representation fo the function call
}
\description{
Constructs a function call from a name or a function and a list of arguments to be passed to it.
}
\examples{
do.call('abs', list(x=-10))
str_call('abs', list(x=-10))
}
\seealso{
\code{\link{do.call}}
}
