% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{template}
\usage{
template(text, ...)
}
\arguments{
\item{text}{character: text in which is replaced}

\item{...}{named list of parameters for replacing}
}
\value{
character: replaced text
}
\description{
Replaces in text the terms \code{\{\{...\}\}} by the parameters given in \code{...}. 
Note that if the parameters contain \code{\{\{...\}\}} then it will \strong{not} be replaced recursively.
}
\examples{
template('{{a}}+1', a='test')
}
