% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textOutput.R
\name{textOutput}
\alias{textOutput}
\title{textOutput}
\usage{
textOutput(outputId, container = if (inline) span else div, inline = FALSE,
  file = NULL, text = NULL)
}
\arguments{
\item{file}{character: name of file with R code to execute}

\item{text}{character: text with R code to execute}
}
\description{
Render a reactive output variable as text within an application page. For the not documented parameters see \code{\link[shiny]{textOutput}}
}
\examples{
textOutput('mytext', text='1:10')
\dontrun{
# create a static app
makeShinyApp(output=textOutput('mytext', text='1:10'))
}
}
