% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNum.R
\name{toInt}
\alias{toInt}
\title{toInt}
\usage{
toInt(x, min = -Inf, max = +Inf)
}
\arguments{
\item{x}{input object}

\item{min}{numeric: minimal value}

\item{max}{numeric: maximal value}
}
\value{
a single integer value
}
\description{
Converts \code{x} to an integer. 
If the conversion fails or the integer is outside \code{min} and \code{max} then \code{NA_integer_} is returned
}
\examples{
toInt(3.0)
toInt("3.0")
toInt("test")
}
