% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNum.R
\name{toLog}
\alias{toLog}
\title{toLog}
\usage{
toLog(x)
}
\arguments{
\item{x}{input object}
}
\value{
a single logical value
}
\description{
Converts \code{x} to a logical. If the conversion fails \code{NA} is returned.
}
\examples{
toLog(3.0)
toLog("T")
toLog("test")
}
