% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNum.R
\name{toNum}
\alias{toNum}
\title{toNum}
\usage{
toNum(x, min = -Inf, max = +Inf)
}
\arguments{
\item{x}{input object}

\item{min}{numeric: minimal value}

\item{max}{numeric: maximal value}
}
\value{
a single integer value
}
\description{
Converts \code{x} to a numeric. 
If the conversion fails or the value is outside \code{min} and \code{max} then \code{NA} is returned
}
\examples{
toNum(3.0)
toNum("3.0")
toNum("test")
}
