% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translations.R
\name{translations}
\alias{translations}
\title{translations}
\usage{
translations(path = ".")
}
\arguments{
\item{path}{character: path to the directory with po files}
}
\value{
returns an object with translations
}
\description{
Loads \code{po} files with translations into a object
}
\examples{
pomessages <- translations()
}
