% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varIn.R
\name{varIn}
\alias{varIn}
\title{varIn}
\usage{
varIn(inputId, label, selected = NULL, selectize = TRUE,
  width = NULL, size = NULL, type = is.true())
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{selected}{The initially selected value (or multiple values if
\code{multiple = TRUE}). If not specified then defaults to the first value
for single-select lists and no values for multiple select lists.}

\item{selectize}{Whether to use \pkg{selectize.js} or not.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}

\item{size}{Number of items to show in the selection box; a larger number
will result in a taller box. Not compatible with \code{selectize=TRUE}.
Normally, when \code{multiple=FALSE}, a select input will be a drop-down
list, but when \code{size} is set, it will be a box instead.}

\item{type}{function: returns for each variable \code{TRUE} or \code{FALSE}}
}
\value{
an environment
}
\description{
varIn
}
\examples{
varIn('var1')
}
