% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.choices.R
\name{as.choices}
\alias{as.choices}
\title{as.choices}
\usage{
as.choices(txt, inc = NULL)
}
\arguments{
\item{txt}{character vector: element names}

\item{inc}{logical vector: if a specific \code{txt} element should be included (default:\code{NULL} means all elements are included)}
}
\value{
a list
}
\description{
Creates a list with named elements according to \code{txt}. 
The list elements are the numbers from 1 to \code{length(txt)}.
}
\examples{
as.choices(c("German", "English"))
as.choices(c("German", "English", "Spanish"), c(TRUE, FALSE, TRUE))
}
