% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkboxIn.R
\name{checkboxIn}
\alias{checkboxIn}
\title{checkboxIn}
\usage{
checkboxIn(inputId, label, value = FALSE, width = NULL)
}
\arguments{
\item{inputId}{character: name of checkbox element}

\item{label}{character: label of input element}

\item{value}{logical: start value (default: \code{FALSE})}

\item{width}{numeric: width of input element}
}
\value{
a checkbox for use into \code{\link{makeShinyApp}}
}
\description{
Creates a checkbox element, for more details see \code{\link[shiny]{checkboxInput}}.
}
\examples{
checkbox('rug', 'Show observations')
}
