% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMake.R
\name{dataMake}
\alias{dataMake}
\title{dataMake}
\usage{
dataMake(file, ...)
}
\arguments{
\item{file}{character: name of RDS file}

\item{...}{list of named(!) data set(s)}
}
\value{
nothing
}
\description{
Creates a list of data sets (only data frames are allowed!) and saves them to a file. These file(s) can be used in \code{\link{dataIn}}.
}
\examples{
dataMake('MA.rds', MotorTrend=mtcars, AndersonIris=iris)
}
