% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{template}
\usage{
template(text = NULL, file = NULL, ...)
}
\arguments{
\item{text}{character: text in which is replaced}

\item{file}{character: name fo file}

\item{...}{list of parameeter for replacing}
}
\value{
character: replaced text
}
\description{
Replaces in text the terms \code{\{\{...\}\}} by the parameters given in \code{...}. Either a \code{text} or \code{file} must be given.
}
\examples{
template(text='{{a}}+1', a='test')
}
